Maria/requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.0.0
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.1
cachetools==5.3.2
catboost==1.2.2
certifi==2023.7.22
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.2.0
Click==7.0
cloudpickle==3.0.0
colorama==0.4.3
colorlog==6.8.0
comm==0.1.4
command-not-found==0.3
contourpy==1.1.1
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
debugpy==1.8.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info==0.23+ubuntu1.1
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
exceptiongroup==1.1.3
executing==1.2.0
fasteners==0.14.1
fastjsonschema==2.18.0
fonttools==4.42.1
frozenlist==1.4.1
future==0.18.2
gitdb==4.0.11
GitPython==3.1.40
graphviz==0.20.1
greenlet==3.0.3
h11==0.14.0
httplib2==0.14.0
idna==2.8
imageio==2.33.1
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipykernel==6.25.2
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.1.1
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-client==8.3.1
jupyter-core==5.3.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter-server==2.7.3
jupyter-server-terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-server==2.25.0
jupyterlab-widgets==3.0.9
keyring==18.0.1
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-loader==0.3
lightgbm==4.2.0
lime==0.2.0.1
llvmlite==0.41.1
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
magic-filter==1.0.12
Mako==1.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mdurl==0.1.2
missingno==0.5.2
mistune==3.0.1
mlxtend==0.23.0
monotonic==1.5
more-itertools==4.2.0
multidict==6.0.4
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
netifaces==0.10.4
networkx==3.1
notebook==7.0.3
notebook-shim==0.2.3
numba==0.58.1
numpy==1.24.4
oauthlib==3.1.0
olefile==0.46
opencv-python==4.8.1.78
optuna==3.5.0
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
patsy==0.5.4
pexpect==4.6.0
phantomjs==1.4.1
pickleshare==0.7.5
Pillow==10.1.0
pkgutil-resolve-name==1.3.10
platformdirs==3.10.0
plotly==5.18.0
plotly-express==0.4.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.25.1
psutil==5.9.5
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
pydantic==2.5.3
pydantic-core==2.14.6
pydeck==0.8.1b0
Pygments==2.16.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.1.1
pyRFC3339==1.1
pyrsistent==0.15.5
PySocks==1.7.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-json-logger==2.0.7
pytz==2023.3.post1
PyWavelets==1.4.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==25.1.1
referencing==0.30.2
reportlab==3.5.34
requests==2.31.0
requests-unixsocket==0.2.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.10.3
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.10.1
seaborn==0.13.0
SecretStorage==2.3.1
selenium==4.14.0
Send2Trash==1.8.2
shap==0.44.0
simplejson==3.16.0
six==1.14.0
slicer==0.0.7
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.25
stack-data==0.6.2
statsmodels==0.14.0
streamlit==1.29.0
systemd-python==234
tenacity==8.2.3
terminado==0.17.1
texttable==1.6.2
threadpoolctl==3.2.0
tifffile==2023.7.10
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
trio==0.22.2
trio-websocket==0.11.1
typing-extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==2.0.7
usb-creator==0.3.7
validators==0.22.0
wadllib==1.3.3
watchdog==3.0.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.6.3
widgetsnbextension==4.0.9
wsproto==1.2.0
xgboost==2.0.3
xkit==0.0.0
yarl==1.9.4
zipp==3.16.2
